type changes in api CODE added

+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/entity/PlantMaintenanceTransaction.java
@@ -38,7 +38,7 @@ public class PlantMaintenanceTransaction {
     private Date maintForMonth;
 
     @Column(name = "Rate")
-    private Double rate;
+    private Float rate;
 
     @Column(name = "Remarks", length = 500)
     private String remarks;
